+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.4.80 2025-04-11
+
+- Automatic Jest Detection and possibility to define Include & Exclude patterns
+- New Vscode Test API & Test Explorer
+- Use NPX to find jest Binrary
+
 ## 0.4.52 - 2022-08-14
 
 - Let CI continously publish extension to ovsx 
@@ -139,7 +145,7 @@
 
 ### Added
 
-- add new settings codeLensSelector which enables CodeLens for files matching this pattern (default **/*.{test,spec}.{js,jsx,ts,tsx})
+- add new settings testFilePattern which enables CodeLens for files matching this pattern (default **/*.{test,spec}.{js,jsx,ts,tsx})
 
 ## 0.4.21 - 2020-06-02